Basic refactor, isolate battlecon components into their own folder.
