+++ sigil/src/components/BattleCon/StatusTerm.jsx
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from "react";
 import '~/styles/StatusTerm.css'
-import battltermIcon from '../assets/ui/battleterm_icon.png'
+import battltermIcon from '~/assets/ui/battleterm_icon.png'
 
 const StatusTerm = ({messages}) => {
